+++ packages/cookies/dev/index.tsx
@@ -1,8 +1,9 @@
-import { Component, createSignal } from "solid-js";
+import { Component } from "solid-js";
+import { createUserTheme } from "../src/index.js";
 
 const App: Component = () => {
-  const [count, setCount] = createSignal(0);
-  const increment = () => setCount(count() + 1);
+  const [theme, setTheme] = createUserTheme();
+  const increment = () => setTheme(theme() === "light" ? "dark" : "light");
   return (
     <div class={`min-h-screen ${"dark" == "dark" ? "dark" : ""}`}>
       <div class="box-border flex min-h-screen w-full flex-col items-center justify-center space-y-4 bg-gray-800 p-24 text-white">
@@ -10,7 +11,7 @@ const App: Component = () => {
           <h4>Counter component</h4>
           <p class="caption">it's very important...</p>
           <button class="btn" onClick={increment}>
-            {count()}
+            {theme()}
           </button>
         </div>
       </div>