+++ packages/resource/src/index.ts
@@ -152,7 +152,7 @@ export function makeCache<T, S, R>(
   const save = () => {
     try {
       options?.storage?.setItem(key, serialize(localCache));
-    } catch (e) {
+    } catch (_) {
       /**/
     }
   };
@@ -163,7 +163,7 @@ export function makeCache<T, S, R>(
   if (options?.storage) {
     try {
       Object.assign(localCache, deserialize(options.storage.getItem(key) || "{}"));
-    } catch (e) {
+    } catch (_) {
       /**/
     }
   }